refactor(client): split schema-initializer items into multiple files (#744)
