+++ packages/core/client/src/schema-initializer/items/ActionInitializer.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import { InitializerWithSwitch } from './InitializerWithSwitch';
+
+export const ActionInitializer = (props) => {
+  return <InitializerWithSwitch {...props} type={'x-action'} />;
+};
