+++ packages/core/client/src/schema-initializer/items/BlockInitializer.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import { SchemaInitializer } from "..";
+
+// Block
+export const BlockInitializer = (props) => {
+  const { item, insert } = props;
+  return (
+    <SchemaInitializer.Item
+      onClick={() => {
+        insert({
+          ...item.schema,
+        });
+      }}
+    />
+  );
+};
