+++ packages/core/client/src/schema-initializer/items/CreateFormBlockInitializer.tsx
@@ -1,0 +1,49 @@
+import React from "react";
+import { FormOutlined } from '@ant-design/icons';
+
+import { useBlockAssociationContext } from "../../block-provider";
+import { useCollection } from "../../collection-manager";
+import { useSchemaTemplateManager } from "../../schema-templates";
+import { SchemaInitializer } from "../SchemaInitializer";
+import { createFormBlockSchema, useRecordCollectionDataSourceItems } from "../utils";
+
+export const CreateFormBlockInitializer = (props) => {
+    const { onCreateBlockSchema, componentType, createBlockSchema, insert, ...others } = props;
+    const { getTemplateSchemaByMode } = useSchemaTemplateManager();
+    const association = useBlockAssociationContext();
+    const collection = useCollection();
+    return (
+      <SchemaInitializer.Item
+        icon={<FormOutlined />}
+        {...others}
+        onClick={async ({ item }) => {
+          if (item.template) {
+            const s = await getTemplateSchemaByMode(item);
+            if (item.template.componentName === 'FormItem') {
+              const blockSchema = createFormBlockSchema({
+                actionInitializers: 'CreateFormActionInitializers',
+                association,
+                collection: collection.name,
+                template: s,
+              });
+              if (item.mode === 'reference') {
+                blockSchema['x-template-key'] = item.template.key;
+              }
+              insert(blockSchema);
+            } else {
+              insert(s);
+            }
+          } else {
+            insert(
+              createFormBlockSchema({
+                actionInitializers: 'CreateFormActionInitializers',
+                association,
+                collection: collection.name,
+              }),
+            );
+          }
+        }}
+        items={useRecordCollectionDataSourceItems('FormItem')}
+      />
+    );
+  };
