+++ packages/core/client/src/schema-initializer/items/CustomizeActionInitializer.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import { BlockInitializer } from '.';
+
+export const CustomizeActionInitializer = (props) => {
+  return <BlockInitializer {...props} />;
+};
