+++ packages/core/client/src/schema-initializer/items/RecordAssociationBlockInitializer.tsx
@@ -1,0 +1,38 @@
+import React from "react";
+import { TableOutlined } from '@ant-design/icons';
+
+import { useCollectionManager } from "../../collection-manager";
+import { useSchemaTemplateManager } from "../../schema-templates";
+import { SchemaInitializer } from "../SchemaInitializer";
+import { createTableBlockSchema, useRecordCollectionDataSourceItems } from "../utils";
+
+export const RecordAssociationBlockInitializer = (props) => {
+  const { item, onCreateBlockSchema, componentType, createBlockSchema, insert, ...others } = props;
+  const { getTemplateSchemaByMode } = useSchemaTemplateManager();
+  const { getCollection } = useCollectionManager();
+  const field = item.field;
+  const collection = getCollection(field.target);
+  const resource = `${field.collectionName}.${field.name}`;
+  return (
+    <SchemaInitializer.Item
+      icon={<TableOutlined />}
+      {...others}
+      onClick={async ({ item }) => {
+        if (item.template) {
+          const s = await getTemplateSchemaByMode(item);
+          insert(s);
+        } else {
+          insert(
+            createTableBlockSchema({
+              rowKey: collection.filterTargetKey,
+              collection: field.target,
+              resource,
+              association: resource,
+            }),
+          );
+        }
+      }}
+      items={useRecordCollectionDataSourceItems('Table', item, field.target, resource)}
+    />
+  );
+};
