+++ packages/core/client/src/schema-initializer/items/TableActionColumnInitializer.tsx
@@ -1,0 +1,27 @@
+import React from "react";
+
+import { InitializerWithSwitch } from "./InitializerWithSwitch";
+
+export const TableActionColumnInitializer = (props) => {
+  const schema = {
+    type: 'void',
+    title: '{{ t("Actions") }}',
+    'x-decorator': 'TableV2.Column.ActionBar',
+    'x-component': 'TableV2.Column',
+    'x-designer': 'TableV2.ActionColumnDesigner',
+    'x-initializer': 'TableActionColumnInitializers',
+    'x-action-column': 'actions',
+    properties: {
+      actions: {
+        type: 'void',
+        'x-decorator': 'DndContext',
+        'x-component': 'Space',
+        'x-component-props': {
+          split: '|',
+        },
+        properties: {},
+      },
+    },
+  };
+  return <InitializerWithSwitch {...props} schema={schema} type={'x-action-column'} />;
+};
