+++ packages/core/client/src/schema-initializer/items/TableBlockInitializer.tsx
@@ -1,0 +1,23 @@
+import React from "react";
+import { TableOutlined } from '@ant-design/icons';
+
+import { useCollectionManager } from "../../collection-manager";
+import { DataBlockInitializer } from "./DataBlockInitializer";
+import { createTableBlockSchema } from "../utils";
+
+export const TableBlockInitializer = (props) => {
+  const { insert } = props;
+  const { getCollection } = useCollectionManager();
+  return (
+    <DataBlockInitializer
+      {...props}
+      icon={<TableOutlined />}
+      componentType={'Table'}
+      onCreateBlockSchema={async ({ item }) => {
+        const collection = getCollection(item.name);
+        const schema = createTableBlockSchema({ collection: item.name, rowKey: collection.filterTargetKey || 'id' });
+        insert(schema);
+      }}
+    />
+  );
+};
