+++ packages/core/client/src/schema-initializer/items/TableSelectorInitializer.tsx
@@ -1,0 +1,29 @@
+import React from "react";
+import { FormOutlined } from '@ant-design/icons';
+
+import { useCollection } from "../../collection-manager";
+import { useSchemaTemplateManager } from "../../schema-templates";
+import { SchemaInitializer } from "../SchemaInitializer";
+import { createTableSelectorSchema } from "../utils";
+
+export const TableSelectorInitializer = (props) => {
+  const { onCreateBlockSchema, componentType, createBlockSchema, insert, ...others } = props;
+  const { getTemplateSchemaByMode } = useSchemaTemplateManager();
+  const collection = useCollection();
+  return (
+    <SchemaInitializer.Item
+      icon={<FormOutlined />}
+      {...others}
+      onClick={async ({ item }) => {
+        const field = item.field;
+        insert(
+          createTableSelectorSchema({
+            rowKey: collection.filterTargetKey,
+            collection: collection.name,
+            resource: collection.name,
+          }),
+        );
+      }}
+    />
+  );
+};
