+++ packages/core/client/src/schema-initializer/items/ViewActionInitializer.tsx
@@ -1,0 +1,52 @@
+import React from "react";
+
+import { ActionInitializer } from "./ActionInitializer";
+
+export const ViewActionInitializer = (props) => {
+    const schema = {
+      type: 'void',
+      title: '{{ t("View") }}',
+      'x-action': 'view',
+      'x-designer': 'Action.Designer',
+      'x-component': 'Action',
+      'x-component-props': {
+        openMode: 'drawer',
+      },
+      properties: {
+        drawer: {
+          type: 'void',
+          title: '{{ t("View record") }}',
+          'x-component': 'Action.Container',
+          'x-component-props': {
+            className: 'nb-action-popup',
+          },
+          properties: {
+            tabs: {
+              type: 'void',
+              'x-component': 'Tabs',
+              'x-component-props': {},
+              'x-initializer': 'TabPaneInitializers',
+              properties: {
+                tab1: {
+                  type: 'void',
+                  title: '{{t("Details")}}',
+                  'x-component': 'Tabs.TabPane',
+                  'x-designer': 'Tabs.Designer',
+                  'x-component-props': {},
+                  properties: {
+                    grid: {
+                      type: 'void',
+                      'x-component': 'Grid',
+                      'x-initializer': 'RecordBlockInitializers',
+                      properties: {},
+                    },
+                  },
+                },
+              },
+            },
+          },
+        },
+      },
+    };
+    return <ActionInitializer {...props} schema={schema} />;
+  };
